Переход на 8-битный вывод экрана

+++ psp/src/filler.h
@@ -38,7 +38,7 @@ private:
     uint32_t pixel32_grouped;
 #endif
     uint32_t * mem32;
-    uint32_t * pixels;
+    uint8_t * pixels;
 
     Memory & memory;
     IO & io;
